zendesk/requirements.txt
google-cloud-pubsub
google-cloud-secret-manager
python-dotenv
requests

# https://arrow.readthedocs.io/en/latest/
arrow